+++ relevanceai/analytics_funcs.py
@@ -1,17 +1,23 @@
 import analytics
 import asyncio
-
+import json
 from typing import Callable
 
 from functools import wraps
 
 from relevanceai.config import CONFIG
 from relevanceai.json_encoder import json_encoder
+from relevanceai.logger import FileLogger
 
 
 def enable_tracking():
     if CONFIG.is_field("mixpanel.enable_tracking", CONFIG.config):
         return CONFIG.get_field("mixpanel.enable_tracking", CONFIG.config)
+
+
+def get_json_size(json_obj):
+    # Returns it in bytes
+    return len(json.dumps(json_obj).encode("utf-8")) / 1024
 
 
 def track(func: Callable):
@@ -22,7 +28,7 @@ def track(func: Callable):
 
                 async def send_analytics():
                     user_id = args[0].firebase_uid
-                    event = f"pysdk-{func.__name__}"
+                    event_name = f"pysdk-{func.__name__}"
                     kwargs.update(dict(zip(func.__code__.co_varnames, args)))
                     properties = {
                         "args": args,
@@ -30,15 +36,25 @@ def track(func: Callable):
                     }
                     if user_id is not None:
                         # Upsert/inserts/updates are too big to track
-                        if "insert" in event or "upsert" in event or "update" in event:
+                        if (
+                            "insert" in event_name
+                            or "upsert" in event_name
+                            or "update" in event_name
+                            or "fit" in event_name
+                            or "predict" in event_name
+                            or get_json_size(
+                                json_encoder(properties, force_string=True)
+                            )
+                            > 30
+                        ):
                             analytics.track(
                                 user_id=user_id,
-                                event=event,
+                                event=event_name,
                             )
                         else:
                             analytics.track(
                                 user_id=user_id,
-                                event=event,
+                                event=event_name,
                                 properties=json_encoder(properties, force_string=True),
                             )