+++ relevanceai/cluster_report/cluster_report.py
@@ -98,7 +98,6 @@ from typing import Union, List, Dict, Any, Optional
 import functools
 from warnings import warn
 from doc_utils import DocUtils
-from relevanceai.analytics_funcs import track
 
 try:
     from sklearn.metrics import (
@@ -145,11 +144,10 @@ class ClusterReport(DocUtils):
         from the model.
     """
 
-    @track
     def __init__(
         self,
         X: Union[list, np.ndarray],
-        cluster_labels: Union[List[Union[str, float]], np.ndarray[Any, Any]],
+        cluster_labels: Union[List[Union[str, float]], np.ndarray],
         model: KMeans = None,
         num_clusters: int = None,
         outlier_label: Union[str, int] = -1,
@@ -288,7 +286,6 @@ class ClusterReport(DocUtils):
             return
 
     @property  # type: ignore
-    @track
     @functools.lru_cache(maxsize=128)
     def internal_report(self):
         """
@@ -300,7 +297,7 @@ class ClusterReport(DocUtils):
         graded_score = self.X_silhouette_scores.mean()
         grade = get_silhouette_grade(graded_score)
 
-        _internal_report = {
+        self._internal_report = {
             "grade": grade,
             "overall": {
                 "summary": ClusterReport.summary_statistics(self.X),
@@ -314,13 +311,21 @@ class ClusterReport(DocUtils):
                     self.X_silhouette_scores
                 ),
             },
-            "each": {
-                "summary": {},
-                "centers": {},
-                "silhouette_score": {},
-            },
+            "each": [
+                # {
+                #     "cluster_id": "cluster-1",
+                #     "summary": {},
+                #     "centers": {},
+                #     "silhouette_score": {}
+                # }
+            ]
+            # {
+            #     "summary": {},
+            #     "centers": {},
+            #     "silhouette_score": {},
+            # },
         }
-        self._store_basic_centroid_stats(_internal_report["overall"])
+        self._store_basic_centroid_stats(self._internal_report["overall"])
 
         labels, counts = np.unique(self.cluster_labels, return_counts=True)
         if self.verbose:
@@ -330,6 +335,9 @@ class ClusterReport(DocUtils):
         cluster_report = {"frequency": {"total": 0, "each": {}}}
 
         for i, cluster_label in enumerate(labels):
+            cluster_label_doc = {
+                "cluster_id": str(cluster_label),
+            }
             cluster_bool = self.cluster_labels == cluster_label
 
             specific_cluster_data = self.X[cluster_bool]
@@ -339,9 +347,7 @@ class ClusterReport(DocUtils):
             cluster_report["frequency"]["total"] += cluster_frequency
             cluster_report["frequency"]["each"][cluster_label] = cluster_frequency
 
-            _internal_report["each"]["summary"][
-                cluster_label
-            ] = ClusterReport.summary_statistics(self.X)
+            cluster_label_doc["summary"] = ClusterReport.summary_statistics(self.X)
 
             # If each value of the vector is important
             center_stats = {"by_features": {}}
@@ -365,7 +371,9 @@ class ClusterReport(DocUtils):
                     )
                 )
 
-                _internal_report["overall"]["grand_centroids"].append(grand_centroid)
+                self._internal_report["overall"]["grand_centroids"].append(
+                    grand_centroid
+                )
 
                 center_stats[
                     "distance_from_centroid"
@@ -399,22 +407,22 @@ class ClusterReport(DocUtils):
                     "overall_z_score"
                 ] = ClusterReport.get_z_score(
                     centroid_vector,
-                    _internal_report["overall"]["summary"]["mean"],
-                    _internal_report["overall"]["summary"]["std"],
+                    self._internal_report["overall"]["summary"]["mean"],
+                    self._internal_report["overall"]["summary"]["std"],
                 )
 
                 center_stats["by_features"]["z_score"] = ClusterReport.get_z_score(
                     centroid_vector,
-                    _internal_report["each"]["summary"][cluster_label]["mean"],
-                    _internal_report["each"]["summary"][cluster_label]["std"],
+                    cluster_label_doc["summary"]["mean"],
+                    cluster_label_doc["summary"]["std"],
                 )
 
                 center_stats["by_features"][
                     "overall_z_score_grand_centroid"
                 ] = ClusterReport.get_z_score(
                     grand_centroid,
-                    _internal_report["overall"]["summary"]["mean"],
-                    _internal_report["overall"]["summary"]["std"],
+                    self._internal_report["overall"]["summary"]["mean"],
+                    self._internal_report["overall"]["summary"]["std"],
                 )
 
                 center_stats[
@@ -424,9 +432,8 @@ class ClusterReport(DocUtils):
                 )
 
                 center_stats["by_features"]["z_score_grand_centroid"] = (
-                    grand_centroid
-                    - _internal_report["each"]["summary"][cluster_label]["mean"]
-                ) / _internal_report["each"]["summary"][cluster_label]["std"]
+                    grand_centroid - cluster_label_doc["summary"]["mean"]
+                ) / cluster_label_doc["summary"]["std"]
 
                 # this might not be needed
                 center_stats["overall_z_score"] = ClusterReport.summary_statistics(
@@ -448,38 +455,44 @@ class ClusterReport(DocUtils):
                     squared_errors, axis=2
                 )
 
-                squared_errors_by_col = {}
+                squared_errors_by_col = []
 
                 for f in range(len(squared_errors[0])):
-                    squared_errors_by_col[f] = ClusterReport.summary_statistics(
-                        squared_errors[:, f], axis=2
+                    squared_errors_by_col.append(
+                        {
+                            "cluster_id": str(f),
+                            "squared_errors": ClusterReport.summary_statistics(
+                                squared_errors[:, f], axis=2
+                            ),
+                        }
                     )
 
-                center_stats["by_features"]["squared_errors"] = squared_errors_by_col
+                center_stats["by_features"] = squared_errors_by_col
 
-                _internal_report["each"]["centers"][cluster_label] = center_stats
+                cluster_label_doc["centers"] = center_stats
 
-            _internal_report["each"]["silhouette_score"][
-                cluster_label
-            ] = ClusterReport.summary_statistics(
+            cluster_label_doc["silhouette_score"] = ClusterReport.summary_statistics(
                 self.X_silhouette_scores[cluster_bool], axis=2
             )
+
+            self._internal_report["each"].append(cluster_label_doc)
 
         if self.has_centers():
 
             min_centroid_distance = min(
-                c["distance_from_centroid"]["min"]
-                for c in _internal_report["each"]["centers"].values()
+                c["centers"]["distance_from_centroid"]["min"]
+                for c in self._internal_report["each"]
             )
 
-            max_centroid_distance = _internal_report["overall"][
+            max_centroid_distance = self._internal_report["overall"][
                 "centroids_distance_matrix"
             ].max()
-            _internal_report["overall"]["dunn_index"] = self.dunn_index(
+
+            self._internal_report["overall"]["dunn_index"] = self.dunn_index(
                 min_centroid_distance, max_centroid_distance
             )
 
-        return _internal_report
+        return self._internal_report
 
     def _get_centroid_vector(
         self, i: int = None, cluster_label: int = None, default_vector=None
@@ -574,7 +587,6 @@ class ClusterReport(DocUtils):
             return pd.concat([metrics, overall_df.reset_index()], axis=1).fillna(" ")
 
     @staticmethod
-    @track
     def calculate_centroids(X, cluster_labels):
         """Calculate the centes"""
         centroid_vectors = {}
@@ -583,7 +595,6 @@ class ClusterReport(DocUtils):
         return centroid_vectors
 
     @staticmethod
-    @track
     def calculate_medoids(X, cluster_labels):
         centroids = ClusterReport.calculate_centroids(X, cluster_labels)
         medoids = {}